feat: 혼잡도 컴포넌트 공통으로 뽑음
